+++ src/rex.mobile/src/rex/mobile/__init__.py
@@ -8,4 +8,5 @@ from .errors import *
 from .initializer import *
 from .interface import *
 from .settings import *
+from .skeleton import *